15/10/22: Added PolicyNotFoundException() to the Policy CRUD

+++ insurance/src/main/java/com/insurance/controller/AdminController.java
@@ -10,17 +10,20 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Optional;
 
+import javax.servlet.http.HttpServletResponse;
+
 import com.insurance.dao.PolicyRepository;
 import com.insurance.dao.UserRepository;
 import com.insurance.entity.Policy;
 import com.insurance.entity.User;
+import com.insurance.exceptions.PolicyNotFoundException;
 
 @RestController
 public class AdminController {
 
 	@Autowired
 	PolicyRepository policyRepo;
-	
+
 	@Autowired
 	UserRepository userRepo;
 
@@ -38,9 +41,9 @@ public class AdminController {
 	@RequestMapping(value = "/admin/policy/{id}", method = RequestMethod.PUT)
 	public ResponseEntity<Policy> updatePolicy(@PathVariable("id") Integer id, @RequestBody Policy policy) {
 		Optional<Policy> optP = policyRepo.findById(id);
-//		if (optP.isEmpty()) {
-//			throw new PolicyNotFoundException();
-//		}
+		if (optP.isEmpty()) {
+			throw new PolicyNotFoundException();
+		}
 
 		Policy p = new Policy();
 		p = optP.get();
@@ -52,46 +55,40 @@ public class AdminController {
 	}
 
 	// ADMIN : List all Policies
-	@RequestMapping(value="/admin/policy", method = RequestMethod.GET)
+	@RequestMapping(value = "/admin/policy", method = RequestMethod.GET)
 	public Iterable<Policy> ListPolicy() {
 		return policyRepo.findAll();
 	}
-	
+
+	// ADMIN : List Policy by ID
+	@RequestMapping(value = "/admin/policy/{id}", method = RequestMethod.GET)
+	public Policy getPolicyById(@PathVariable("id") Integer id) {
+		Optional<Policy> optP = policyRepo.findById(id);
+		if(optP.isEmpty()) {
+			throw new PolicyNotFoundException();
+		}
+		return policyRepo.findById(id).get();
+	}
+
+
 	// ADMIN : Delete a Policy
-	@RequestMapping(value="/admin/policy/{id}", method=RequestMethod.DELETE)
+	@RequestMapping(value = "/admin/policy/{id}", method = RequestMethod.DELETE)
 	public String deletePolicy(@PathVariable("id") Integer id) {
 		Optional<Policy> optP = policyRepo.findById(id);
-//		if (optP.isEmpty()) {
-////			return "Invalid username or Password.";
-//			throw new PolicyNotFoundException();
-//		}
+		if (optP.isEmpty()) {
+			throw new PolicyNotFoundException();
+		}
 		Policy p = new Policy();
 		p = optP.get();
-		
-		
-		
-		//CODE TO REMOVE policyNum FROM THE RESPECTIVE USER
-		
-		// error
-		// creating new user instead of editing existing
-		
-//		// getting User with this policy
-//		int uId=p.getUserId();
-//		
-//		// removing the policyNum from user entity
-//		User u = new User();
-//		Optional<User> optU = userRepo.findById(uId);
-//		u.setPolicyNum(null);
-//		final User updatedUser = userRepo.save(u);
-////		return ResponseEntity.ok(updatedUser);
-	
-		
-		
-		//delete the Policy
-		policyRepo.deleteById(id);
-		
-		return "Policy Deleted.";
-		
+
+		// delete the Policy if not in use
+		if (p.getUserId() == 0) {
+			policyRepo.deleteById(id);
+
+			return "Policy Deleted.";
+		} else {
+			return "Policy in use, cannot be deleted.";
+		}
 	}
-	
+
 }
